debate_day/agents/mod/requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0 